Update
Fron-end, function

+++ pages/actions/post_handle.php
@@ -22,8 +22,8 @@
                 $db->close();
     }
 
- /*   if(isset($_POST['post_btn_delete'])){
-            $post_id = $_GET['post_id']
+    if(isset($_POST['post_btn_delete'])){
+            $post_id = $_GET['post_id'];
             //Delete query
             $stmt_dlt = $db->prepare("DELETE FROM post WHERE post_id = $post_id");
 
@@ -31,8 +31,8 @@
             if  (($stmt_dlt->execute()) === TRUE){
                 $_SESSION["dbAddedSuccess"]= true;
                 header("Location: ../classroom_stream.php?class_id=$classid");
-                } else {   
-                echo $stmt_dlt->error;  
+                } else {
+                echo $stmt_dlt->error;
                 }
                 $stmt_dlt->close();
                 $db->close();
@@ -40,11 +40,11 @@
 
 
     if(isset($_POST['post_btn_update'])){
-            $post_id = $_GET['post_id']
+            $post_id = $_GET['post_id'];
             //Update query
             $contentUpdate = mysqli_real_escape_string($db, $_POST["comment_update"]);
             $stmt_upd = $db->prepare("UPDATE post SET content = ?, dateT_update = ? WHERE post_id = ?");
-            $dateTUpdate = date('d/m/Y h:i:s ', time()); 
+            $dateTUpdate = date('d/m/Y h:i:s ', time());
             //another way to set current time: SET dateT_update = now();
             $stmt_upd->bind_param("sss", $contentUpdate,$dateTUpdate,$post_id);
 
@@ -52,10 +52,10 @@
             if  (($stmt_upd->execute()) === TRUE){
                 $_SESSION["dbAddedSuccess"]= true;
                 header("Location: ../classroom_stream.php?class_id=$classid");
-                } else {   
-                echo $stmt_upd->error;  
+                } else {
+                echo $stmt_upd->error;
                 }
                 $stmt_upd->close();
                 $db->close();
-    } */
+    }
 ?>
